Factor out a testing_tools.adapter.pytest package.
