+++ pythonFiles/testing_tools/adapter/pytest/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+from __future__ import absolute_import
+
+from ._cli import add_subparser as add_cli_subparser
+from ._discovery import discover
