+++ pythonFiles/tests/testing_tools/adapter/pytest/test_discovery.py
@@ -1,5 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
+
 from __future__ import print_function, unicode_literals
 
 try:
@@ -11,32 +12,12 @@ import os.path
 import sys
 import unittest
 
-from ...util import Stub, StubProxy
-from testing_tools.adapter.errors import UnsupportedCommandError
+from ....util import Stub, StubProxy
 from testing_tools.adapter.info import TestInfo, TestPath, ParentInfo
-from testing_tools.adapter.pytest import (
-        discover, add_cli_subparser, TestCollector, DiscoveredTests
+from testing_tools.adapter.pytest._discovery import (
+        discover, TestCollector, DiscoveredTests
         )
 import pytest
-
-
-class StubSubparsers(StubProxy):
-
-    def __init__(self, stub=None, name='subparsers'):
-        super(StubSubparsers, self).__init__(stub, name)
-
-    def add_parser(self, name):
-        self.add_call('add_parser', None, {'name': name})
-        return self.return_add_parser
-
-
-class StubArgParser(StubProxy):
-
-    def __init__(self, stub=None):
-        super(StubArgParser, self).__init__(stub, 'argparser')
-
-    def add_argument(self, *args, **kwargs):
-        self.add_call('add_argument', args, kwargs)
 
 
 class StubPyTest(StubProxy):
@@ -166,37 +147,6 @@ class StubPytestConfig(StubProxy):
 
 ##################################
 # tests
-
-class AddCLISubparserTests(unittest.TestCase):
-
-    def test_discover(self):
-        stub = Stub()
-        subparsers = StubSubparsers(stub)
-        parser = StubArgParser(stub)
-        subparsers.return_add_parser = parser
-
-        add_cli_subparser('discover', 'pytest', subparsers)
-
-        self.assertEqual(stub.calls, [
-            ('subparsers.add_parser', None, {'name': 'pytest'}),
-            ])
-
-    def test_unsupported_command(self):
-        subparsers = StubSubparsers(name=None)
-        subparsers.return_add_parser = None
-
-        with self.assertRaises(UnsupportedCommandError):
-            add_cli_subparser('run', 'pytest', subparsers)
-        with self.assertRaises(UnsupportedCommandError):
-            add_cli_subparser('debug', 'pytest', subparsers)
-        with self.assertRaises(UnsupportedCommandError):
-            add_cli_subparser('???', 'pytest', subparsers)
-        self.assertEqual(subparsers.calls, [
-            ('add_parser', None, {'name': 'pytest'}),
-            ('add_parser', None, {'name': 'pytest'}),
-            ('add_parser', None, {'name': 'pytest'}),
-            ])
-
 
 class DiscoverTests(unittest.TestCase):
